Feat: [Client][Components/Captures] フルスクリーン表示の改善
